+++ apps/www/next.config.js
@@ -34,11 +34,6 @@ const nextConfig = {
                 permanent: true,
             },
             {
-                source: "/articles",
-                destination: "/articles/api-recipes",
-                permanent: false,
-            },
-            {
                 source: "/contact",
                 destination: `/contributors/${encodeURIComponent(
                     process.env.NEXT_PUBLIC_CONTACT_CONTRIBUTOR_UUID,